requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
python-dotenv>=1.0
psycopg2-binary>=2.9
gunicorn>=21.2
whitenoise>=6.6
django-cors-headers>=4.3
